updating

+++ dtrainer.py
@@ -221,7 +221,7 @@ def train(model, optimizer, n_epochs, traindataloader, testdataloader=None):
         for batch_idx, inputs in enumerate(traindataloader):
             # input()
             optimizer.zero_grad()
-            with torch.autocast(device_type='cuda', dtype=torch.float32, enabled=use_amp):
+            with torch.autocast(device_type='cuda', dtype=torch.bfloat16, enabled=use_amp):
                 loss, logs = model.training_step(inputs, batch_idx)
             # loss, logs = model.training_step(inputs, batch_idx)
